Updated documentation with new "Developer Notes"
This adds a new "Developer Notes" section, intended for developers and advanced users. This section includes three notes:

- TFRecords: detailed discussion of the TFRecord format with examples of how to read, write, and inspect tfrecords.
- Dataloaders: detailed discussion of Tensorflow and PyTorch dataloaders and sampling methods.
- Custom extractors: discussion of how to create a custom feature extractor with either PyTorch of Tensorflow models.
