+++ slideflow/model/extractors/ctranspath.py
@@ -11,8 +11,6 @@ import torch
 import torch.nn as nn
 import math
 import torch.utils.checkpoint as checkpoint
-import slideflow as sf
-import numpy as np
 from typing import Optional
 from torchvision import transforms
 from huggingface_hub import hf_hub_download
@@ -25,8 +23,7 @@ from timm.models.layers import _assert
 from timm.models.swin_transformer import window_partition, window_reverse
 from timm.models.vision_transformer import checkpoint_filter_fn
 
-from ..base import BaseFeatureExtractor
-from ._slide import features_from_slide
+from ._factory_torch import TorchFeatureExtractor
 
 # -----------------------------------------------------------------------------
 
@@ -559,7 +556,7 @@ class ConvStem(torch.nn.Module):
 
 # -----------------------------------------------------------------------------
 
-class CTransPathFeatures(BaseFeatureExtractor):
+class CTransPathFeatures(TorchFeatureExtractor):
     """
     CTransPath pretrained feature extractor.
 
@@ -589,7 +586,7 @@ class CTransPathFeatures(BaseFeatureExtractor):
 """
 
     def __init__(self, device=None, center_crop=False):
-        super().__init__(backend='torch')
+        super().__init__()
 
         from slideflow.model import torch_utils
         self.device = torch_utils.get_device(device)
@@ -607,7 +604,6 @@ class CTransPathFeatures(BaseFeatureExtractor):
 
         # ---------------------------------------------------------------------
         self.num_features = 768
-        self.num_classes = 0
         all_transforms = [transforms.CenterCrop(224)] if center_crop else []
         all_transforms += [
             transforms.Lambda(lambda x: x / 255.),
@@ -619,20 +615,6 @@ class CTransPathFeatures(BaseFeatureExtractor):
         self.preprocess_kwargs = dict(standardize=False)
         self._center_crop = center_crop
         # ---------------------------------------------------------------------
-
-    def __call__(self, obj, **kwargs):
-        """Generate features for a batch of images or a WSI."""
-        if isinstance(obj, sf.WSI):
-            grid = features_from_slide(self, obj, **kwargs)
-            return np.ma.masked_where(grid == -99, grid)
-        elif kwargs:
-            raise ValueError(
-                f"{self.__class__.__name__} does not accept keyword arguments "
-                "when extracting features from a batch of images."
-            )
-        assert obj.dtype == torch.uint8
-        obj = self.transform(obj)
-        return self.model(obj)
 
     def dump_config(self):
         """Return a dictionary of configuration parameters.
