+++ slideflow/model/extractors/dinov2.py
@@ -1,20 +1,17 @@
 import torch
-import slideflow as sf
-import numpy as np
 
 from torchvision import transforms
 from omegaconf import OmegaConf
 
-import slideflow.dinov2
-from slideflow.dinov2.eval.setup import build_model_for_eval
+from dinov2.eval.setup import build_model_for_eval
+from slideflow.model import torch_utils
 
-from ..base import BaseFeatureExtractor
-from ._slide import features_from_slide
+from ._factory_torch import TorchFeatureExtractor
 
 # -----------------------------------------------------------------------------
 
 
-class DinoV2Features(BaseFeatureExtractor):
+class DinoV2Features(TorchFeatureExtractor):
     """DinoV2 feature extractor.
 
     Feature dimensions: 1024
@@ -34,19 +31,17 @@ class DinoV2Features(BaseFeatureExtractor):
 """
 
     def __init__(self, cfg, weights, device=None, center_crop=False):
-        super().__init__(backend='torch')
+        super().__init__()
 
-        from slideflow.model import torch_utils
         self.cfg = cfg
         self.weights = weights
         self.device = torch_utils.get_device(device)
         self.model = build_model_for_eval(OmegaConf.load(cfg), weights)
-        self.model.fc = torch.nn.Identity().to(self.device)
         self.model.to(self.device)
+        self.model.eval()
 
         # ---------------------------------------------------------------------
         self.num_features = 1024
-        self.num_classes = 0
         all_transforms = [transforms.CenterCrop(224)] if center_crop else []
         all_transforms += [
             transforms.Lambda(lambda x: x / 255.),
@@ -58,20 +53,6 @@ class DinoV2Features(BaseFeatureExtractor):
         self.preprocess_kwargs = dict(standardize=False)
         self._center_crop = center_crop
         # ---------------------------------------------------------------------
-
-    def __call__(self, obj, **kwargs):
-        """Generate features for a batch of images or a WSI."""
-        if isinstance(obj, sf.WSI):
-            grid = features_from_slide(self, obj, **kwargs)
-            return np.ma.masked_where(grid == -99, grid)
-        elif kwargs:
-            raise ValueError(
-                f"{self.__class__.__name__} does not accept keyword arguments "
-                "when extracting features from a batch of images."
-            )
-        assert obj.dtype == torch.uint8
-        obj = self.transform(obj)
-        return self.model(obj)
 
     def dump_config(self):
         """Return a dictionary of configuration parameters.
