+++ slideflow/model/extractors/plip.py
@@ -1,6 +1,4 @@
 import torch
-import slideflow as sf
-import numpy as np
 from torchvision import transforms
 
 try:
@@ -12,8 +10,7 @@ except ImportError:
         "You can install it with 'pip install transformers'."
     )
 
-from ..base import BaseFeatureExtractor
-from ._slide import features_from_slide
+from ._factory_torch import TorchFeatureExtractor
 
 # -----------------------------------------------------------------------------
 
@@ -28,7 +25,7 @@ class CLIPImageFeatures(torch.nn.Module):
         return x
 
 
-class PLIPFeatures(BaseFeatureExtractor):
+class PLIPFeatures(TorchFeatureExtractor):
     """
     PLIP pretrained feature extractor.
     Feature dimensions: 512
@@ -49,7 +46,7 @@ class PLIPFeatures(BaseFeatureExtractor):
 """
 
     def __init__(self, device=None, center_crop=False):
-        super().__init__(backend='torch')
+        super().__init__()
 
         from slideflow.model import torch_utils
         self.device = torch_utils.get_device(device)
@@ -59,7 +56,6 @@ class PLIPFeatures(BaseFeatureExtractor):
 
         # ---------------------------------------------------------------------
         self.num_features = 512
-        self.num_classes = 0
         all_transforms = [transforms.CenterCrop(224)] if center_crop else []
         all_transforms += [
             transforms.Lambda(lambda x: x / 255.),
@@ -71,20 +67,6 @@ class PLIPFeatures(BaseFeatureExtractor):
         self.preprocess_kwargs = dict(standardize=False)
         self._center_crop = center_crop
         # ---------------------------------------------------------------------
-
-    def __call__(self, obj, **kwargs):
-        """Generate features for a batch of images or a WSI."""
-        if isinstance(obj, sf.WSI):
-            grid = features_from_slide(self, obj, **kwargs)
-            return np.ma.masked_where(grid == -99, grid)
-        elif kwargs:
-            raise ValueError(
-                f"{self.__class__.__name__} does not accept keyword arguments "
-                "when extracting features from a batch of images."
-            )
-        assert obj.dtype == torch.uint8
-        obj = self.transform(obj)
-        return self.model(obj)
 
     def dump_config(self):
         """Return a dictionary of configuration parameters.
